Initial commit
Implemented the basic tic-tac-toe tutorial, and added a few features.
The last feature I implemented was the ability to change the sort order.
I also extracted Square and Board into their own files. I feel like the
main index.js file is difficult to read and violates SRP, so I would
like to refactor it and separate out the concerns.